Add filter(Category), Date formatting

+++ server/src/resolverMap.ts
@@ -14,10 +14,23 @@ const resolverMap: IResolvers = {
             distance, checked, currentX, currentY }) => {
             console.debug("Query: getFilteredPlace");
             const earthR = 6371;
+            
+            const CATEGORY = ["한식", "양식", "중식", "일식", "분식"];
 
+            let result;
             // 1. query mongodb with category
-            // cons categoryResult = await Places.find({p_category: ?})
-            const result = await Places.find();
+            if(category.includes("기타")){
+                const filteredCategory = CATEGORY.filter((cate, index) => {
+                    if(category.includes(cate))
+                        return false;
+                    else
+                        return true; 
+                })
+
+                result = await Places.find({ p_category: { $nin: filteredCategory }});
+            }
+            else 
+                result = await Places.find({ p_category: { $in: category } });
             
             // 2. select palce to calculate distance
             const distanceResult = result.filter(place => {
@@ -43,9 +56,6 @@ const resolverMap: IResolvers = {
 
                     return Number(price) >= Number(iPrice)  
                 })
-                
-                //console.log(filteredMenu);
-                //console.log(filteredMenu.length);
 
                 // 가격에 맞는 메뉴가 없으면 지움
                 if(!filteredMenu.length)
@@ -60,7 +70,7 @@ const resolverMap: IResolvers = {
             return dmResult;
         },
         getPlace: async (_:void, { id }) => {
-            console.debug("Query: getPlace");
+            console.log("Query: getPlace");
             const result = await Places.findOne({ p_id: id });
 
             console.log(result);
@@ -70,20 +80,8 @@ const resolverMap: IResolvers = {
          * getPosts Query
          */
         getPosts: async(_: void, args: void) => {
-            console.debug("Query: getPosts");
+            console.log("Query: getPosts");
             const result = await Posts.find();
-            
-            result.forEach(post => {
-                const date = new Date(post.get("createdAt"));
-                post.set("createdAt",  
-                    date.getFullYear() + "-" 
-                            + (date.getMonth()+1 >= 10? date.getMonth()+1 : "0"+(date.getMonth()+1)) + "-"
-                            + (date.getDay() >= 10? date.getDay() : "0"+date.getDay()) + " " 
-                            + (date.getHours() >= 10? date.getHours() : "0"+date.getHours()) + ":"
-                            + (date.getMinutes() >= 10? date.getMinutes() : "0"+date.getMinutes()) + ":" 
-                            + (date.getSeconds() >= 10? date.getSeconds() : "0"+date.getSeconds())
-                , String);
-            })
 
             return result;
         },
@@ -96,48 +94,48 @@ const resolverMap: IResolvers = {
             const updateResult = await Posts.findOneAndUpdate({ _id: id}, { c_views: result!.get("c_views") + 1}, {new: true});
 
             const date = new Date(updateResult!.get("createdAt"));
-            updateResult!.set("createdAt",  
-                    date.getFullYear() + "-" 
-                            + (date.getMonth()+1 >= 10? date.getMonth()+1 : "0"+(date.getMonth()+1)) + "-"
-                            + (date.getDay() >= 10? date.getDay() : "0"+date.getDay()) + " " 
-                            + (date.getHours() >= 10? date.getHours() : "0"+date.getHours()) + ":"
-                            + (date.getMinutes() >= 10? date.getMinutes() : "0"+date.getMinutes()) + ":" 
-                            + (date.getSeconds() >= 10? date.getSeconds() : "0"+date.getSeconds())
-                            , String);
 
             return updateResult;
         }
     },
     Mutation: {
         createPost: async (root, args) => {
-            console.debug("Mutation: createPost");
+            console.log("Mutation: createPost");
 
             const Post = new Posts({
                 c_title: args.title,
                 c_content: args.content,
+                c_author: args.author,
             });
             
             const result = await Post.save();
             console.log(result);
             return result;
         },
-        updatePost: async (root, { id, title, content }) => {
-            console.debug("Mutation: updatePost");
+        updatePost: async (root, args) => {
+            console.log("Mutation: updatePost");
             
-            const updateResult = await Posts.findOneAndUpdate({ _id: id}, { c_title: title, c_content: content }, {new: true})
+            const updateResult = await Posts.findOneAndUpdate({ _id: args.id, c_author: args.author}, { c_title: args.title, c_content: args.content }, {new: true})
 
-            return updateResult; 
+            console.log(updateResult)
+            if(updateResult)
+                return { result: true };
+            else 
+                return { result: false };
         },
-        deletePost: async (root, { id }) => {
-            console.debug("Mutation: deietePost");
+        deletePost: async (root, args) => {
+            console.log("Mutation: deietePost");
 
-            const result = await Posts.findOneAndDelete({ _id: id });
-            console.log(result);
-
-            return { result: true };
+            const result = await Posts.findOneAndDelete({ _id: args.id, c_author: args.author });
+            
+            console.log
+            if(result)
+                return { result: true };
+            else 
+                return { result: false };
         },
         addLikes: async (root, { id }) => {
-            console.debug("Mutation: addLikes");
+            console.log("Mutation: addLikes");
 
             const result = await Posts.findOne({ _id: id });
             const likeResult = await Posts.findOneAndUpdate({ _id: id}, { c_likes: result!.get("c_likes") + 1}, {new: true})
